refactor UserList.js to include user inviting into modal
